// generated/src/app/api-models/prov-credential.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProvCredential {
  seqProvCredential: number;
  seqProvId: number;
  license1: string;
  state1: string;
  license1ExpireDt: Date;
  license2: string;
  state2: string;
  license2ExpireDt: Date;
  deaLicense: string;
  deaExpireDate: Date;
  insuranceCarrier1: string;
  insurancePolicy1: string;
  effectiveFrom1: Date;
  effectiveTo1: Date;
  claimLimit1: number;
  aggregLimit1: number;
  policyDescr1: string;
  insuranceCarrier2: string;
  insurancePolicy2: string;
  effectiveFrom2: Date;
  effectiveTo2: Date;
  claimLimit2: number;
  aggregateLimit2: number;
  policyDescr2: string;
  userDefined1: string;
  userDefined2: string;
  userDefined3: string;
  userDefined4: string;
  userDefined5: string;
  userDefined6: string;
  userDefined7: string;
  userDefined8: string;
  comments: string;
  medicalSchool: string;
  medSchoolDate: Date;
  internship: string;
  internshipDate: Date;
  residency: string;
  residencyDate: Date;
  specialtyBoard1: string;
  board1FromDate: Date;
  board1ToDate: Date;
  specialtyBoard2: string;
  board2FromDate: Date;
  board2ToDate: Date;
  includeInDir: string;
  lastPrintDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  country1: string;
  country2: string;
}